Prometheus reworked & version updated from 0.8.0 to 0.8.1
